Update index.js
Completely rewrites the way, it generates sudoku puzzles. Now it takes in a pre-made "template" from an object/database (see bottom), and randomly assigns numbers to it, then randomly shuffles it while maintaining a solution.
To add more diversity to the puzzles, you could add more templates to the object.
Difficulties are added with this change.